refactor(ff-format): split stream.rs into stream/ subdirectory
Replace the 1315-line monolithic stream.rs with a stream/ subdirectory
containing one file per stream type:

- stream/mod.rs      — re-exports all six public types
- stream/video.rs    — VideoStreamInfo + VideoStreamInfoBuilder + tests
- stream/audio.rs    — AudioStreamInfo + AudioStreamInfoBuilder + tests
- stream/subtitle.rs — SubtitleStreamInfo + SubtitleStreamInfoBuilder + tests

lib.rs is unchanged — pub mod stream; resolves to the directory
automatically. Public API paths are preserved.

Resolves #710
